Shifted gerbers to match offset from working gerbers from NextPCB to see any differences.
I note that the copper was extended to the edge of the board making it look bigger, and that the USB-C charger port at the top right had its vias shrank to avoid shorts due to insufficient space between existing traces/vias. These changes are not updated in KiCAD PCB and likely should be.